requirements.txt
pandas
requests
pymongo
jsonschema
azure-cosmos
